refactor: split files structure
this will reduce the number of line per files and allow us to understand
what is happening (also what has been tested or not)
