+++ internal/controller/maintenance/window_test.go
@@ -1,4 +1,4 @@
-package controller
+package maintenance
 
 import (
 	"testing"
@@ -20,7 +20,7 @@ func spec(windows ...v1alpha1.WindowSpec) *v1alpha1.MaintenanceSpec {
 	return &v1alpha1.MaintenanceSpec{Windows: windows}
 }
 
-func TestCheckMaintenanceWindow_NilAndEmpty(t *testing.T) {
+func TestCheckWindow_NilAndEmpty(t *testing.T) {
 	for _, tc := range []struct {
 		name string
 		spec *v1alpha1.MaintenanceSpec
@@ -29,7 +29,7 @@ func TestCheckMaintenanceWindow_NilAndEmpty(t *testing.T) {
 		{"empty windows", &v1alpha1.MaintenanceSpec{}},
 	} {
 		t.Run(tc.name, func(t *testing.T) {
-			result, err := CheckMaintenanceWindow(tc.spec, time.Now())
+			result, err := CheckWindow(tc.spec, time.Now())
 			if err != nil {
 				t.Fatalf("unexpected error: %v", err)
 			}
@@ -40,7 +40,7 @@ func TestCheckMaintenanceWindow_NilAndEmpty(t *testing.T) {
 	}
 }
 
-func TestCheckMaintenanceWindow_Boundaries(t *testing.T) {
+func TestCheckWindow_Boundaries(t *testing.T) {
 	// Daily window 02:00-06:00 UTC
 	s := spec(window("0 2 * * *", 4*time.Hour, "UTC"))
 
@@ -57,7 +57,7 @@ func TestCheckMaintenanceWindow_Boundaries(t *testing.T) {
 
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {
-			result, err := CheckMaintenanceWindow(s, tc.now)
+			result, err := CheckWindow(s, tc.now)
 			if err != nil {
 				t.Fatalf("unexpected error: %v", err)
 			}
@@ -68,11 +68,11 @@ func TestCheckMaintenanceWindow_Boundaries(t *testing.T) {
 	}
 }
 
-func TestCheckMaintenanceWindow_ReportsWindowEnd(t *testing.T) {
+func TestCheckWindow_ReportsWindowEnd(t *testing.T) {
 	s := spec(window("0 2 * * *", 4*time.Hour, "UTC"))
 	now := time.Date(2025, 6, 15, 3, 0, 0, 0, time.UTC)
 
-	result, _ := CheckMaintenanceWindow(s, now)
+	result, _ := CheckWindow(s, now)
 	if result.CurrentWindowEnd == nil {
 		t.Fatal("expected CurrentWindowEnd to be set")
 	}
@@ -82,11 +82,11 @@ func TestCheckMaintenanceWindow_ReportsWindowEnd(t *testing.T) {
 	}
 }
 
-func TestCheckMaintenanceWindow_ReportsNextWindowStart(t *testing.T) {
+func TestCheckWindow_ReportsNextWindowStart(t *testing.T) {
 	s := spec(window("0 2 * * *", 4*time.Hour, "UTC"))
 	now := time.Date(2025, 6, 15, 10, 0, 0, 0, time.UTC)
 
-	result, _ := CheckMaintenanceWindow(s, now)
+	result, _ := CheckWindow(s, now)
 	if result.NextWindowStart == nil {
 		t.Fatal("expected NextWindowStart to be set")
 	}
@@ -96,7 +96,7 @@ func TestCheckMaintenanceWindow_ReportsNextWindowStart(t *testing.T) {
 	}
 }
 
-func TestCheckMaintenanceWindow_MultipleWindowsUnion(t *testing.T) {
+func TestCheckWindow_MultipleWindowsUnion(t *testing.T) {
 	s := spec(
 		window("0 2 * * *", 4*time.Hour, "UTC"),
 		window("0 20 * * *", 2*time.Hour, "UTC"),
@@ -104,7 +104,7 @@ func TestCheckMaintenanceWindow_MultipleWindowsUnion(t *testing.T) {
 
 	// Inside second window only — union allows it
 	now := time.Date(2025, 6, 15, 21, 0, 0, 0, time.UTC)
-	result, err := CheckMaintenanceWindow(s, now)
+	result, err := CheckWindow(s, now)
 	if err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	}
@@ -114,7 +114,7 @@ func TestCheckMaintenanceWindow_MultipleWindowsUnion(t *testing.T) {
 
 	// Outside both — reports earliest next
 	now = time.Date(2025, 6, 15, 10, 0, 0, 0, time.UTC)
-	result, _ = CheckMaintenanceWindow(s, now)
+	result, _ = CheckWindow(s, now)
 	if result.Allowed {
 		t.Fatal("expected not allowed outside all windows")
 	}
@@ -124,13 +124,13 @@ func TestCheckMaintenanceWindow_MultipleWindowsUnion(t *testing.T) {
 	}
 }
 
-func TestCheckMaintenanceWindow_Timezone(t *testing.T) {
+func TestCheckWindow_Timezone(t *testing.T) {
 	// 2:00 AM Paris time, 4h window
 	s := spec(window("0 2 * * *", 4*time.Hour, "Europe/Paris"))
 
 	// 2:30 AM Paris = 0:30 UTC (CEST +2) — inside
 	now := time.Date(2025, 6, 15, 0, 30, 0, 0, time.UTC)
-	result, err := CheckMaintenanceWindow(s, now)
+	result, err := CheckWindow(s, now)
 	if err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	}
@@ -140,18 +140,18 @@ func TestCheckMaintenanceWindow_Timezone(t *testing.T) {
 
 	// 7:00 AM Paris = 5:00 UTC — outside (window ends 6:00 AM Paris = 4:00 UTC)
 	now = time.Date(2025, 6, 15, 5, 0, 0, 0, time.UTC)
-	result, _ = CheckMaintenanceWindow(s, now)
+	result, _ = CheckWindow(s, now)
 	if result.Allowed {
 		t.Fatal("expected not allowed at 7:00 AM Paris")
 	}
 }
 
-func TestCheckMaintenanceWindow_SpansMidnight(t *testing.T) {
+func TestCheckWindow_SpansMidnight(t *testing.T) {
 	// 22:00 UTC, 6h duration → ends 04:00 next day
 	s := spec(window("0 22 * * *", 6*time.Hour, "UTC"))
 
 	now := time.Date(2025, 6, 16, 1, 0, 0, 0, time.UTC)
-	result, err := CheckMaintenanceWindow(s, now)
+	result, err := CheckWindow(s, now)
 	if err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	}
@@ -160,7 +160,7 @@ func TestCheckMaintenanceWindow_SpansMidnight(t *testing.T) {
 	}
 }
 
-func TestCheckMaintenanceWindow_InvalidInputs(t *testing.T) {
+func TestCheckWindow_InvalidInputs(t *testing.T) {
 	tests := []struct {
 		name string
 		spec *v1alpha1.MaintenanceSpec
@@ -170,7 +170,7 @@ func TestCheckMaintenanceWindow_InvalidInputs(t *testing.T) {
 	}
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {
-			_, err := CheckMaintenanceWindow(tc.spec, time.Now())
+			_, err := CheckWindow(tc.spec, time.Now())
 			if err == nil {
 				t.Fatal("expected error")
 			}
