--- internal/controller/util.go
@@ -1,114 +1,0 @@
-package controller
-
-import (
-	"context"
-	"fmt"
-	"time"
-
-	"github.com/google/uuid"
-	"github.com/home-operations/tuppr/internal/constants"
-	corev1 "k8s.io/api/core/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	tupprv1alpha1 "github.com/home-operations/tuppr/api/v1alpha1"
-)
-
-type Now interface {
-	Now() time.Time
-}
-
-type realClock struct{}
-
-func (r realClock) Now() time.Time {
-	return time.Now()
-}
-
-// GetNodeIp retrieves the InternalIP or ExternalIP of a Kubernetes node while preferring the InternalIP
-func GetNodeIP(node *corev1.Node) (string, error) {
-	externalIP := ""
-
-	for _, addr := range node.Status.Addresses {
-		switch addr.Type {
-		case corev1.NodeInternalIP:
-			return addr.Address, nil
-		case corev1.NodeExternalIP:
-			externalIP = addr.Address
-		}
-	}
-	if externalIP != "" {
-		return externalIP, nil
-	}
-
-	return "", fmt.Errorf("no InternalIP or ExternalIP found for node %q", node.Name)
-}
-
-// GenerateSafeJobName generates a safe Kubernetes Job name
-// It ensures the name is within Kubernetes' 63 character limit
-// by truncating the identifier if necessary and appending a unique suffix.
-func GenerateSafeJobName(prefix, identifier string) string {
-	jobID := uuid.New().String()[:8]
-	const maxLength = 57
-
-	prefixLen := len(prefix) + len(jobID) + 2
-	available := maxLength - prefixLen
-
-	if available <= 0 {
-		return fmt.Sprintf("%s-%s", prefix, jobID)
-	}
-
-	identifierLen := min(len(identifier), available)
-	if identifierLen > 0 {
-		return fmt.Sprintf("%s-%s-%s", prefix, identifier[:identifierLen], jobID)
-	}
-
-	return fmt.Sprintf("%s-%s", prefix, jobID)
-}
-
-// IsAnotherUpgradeActive checks if any other type of upgrade is currently active
-func IsAnotherUpgradeActive(ctx context.Context, c client.Client, currentUpgradeName string, currentUpgradeType string) (bool, string, error) {
-	if currentUpgradeType == "talos" {
-		// Check if any KubernetesUpgrade is InProgress
-		kubernetesUpgrades := &tupprv1alpha1.KubernetesUpgradeList{}
-		if err := c.List(ctx, kubernetesUpgrades); err != nil {
-			return false, "", fmt.Errorf("failed to list KubernetesUpgrade resources: %w", err)
-		}
-
-		for _, upgrade := range kubernetesUpgrades.Items {
-			if upgrade.Status.Phase == constants.PhaseInProgress {
-				return true, fmt.Sprintf("Waiting for KubernetesUpgrade '%s' to complete", upgrade.Name), nil
-			}
-		}
-	} else {
-		// Check if any TalosUpgrade is InProgress
-		talosUpgrades := &tupprv1alpha1.TalosUpgradeList{}
-		if err := c.List(ctx, talosUpgrades); err != nil {
-			return false, "", fmt.Errorf("failed to list TalosUpgrade resources: %w", err)
-		}
-
-		for _, upgrade := range talosUpgrades.Items {
-			if upgrade.Status.Phase == constants.PhaseInProgress || upgrade.Status.Phase == constants.PhasePending {
-				return true, fmt.Sprintf("Waiting for TalosUpgrade '%s' to complete", upgrade.Name), nil
-			}
-		}
-	}
-	if currentUpgradeType == "talos" {
-		talosUpgrades := &tupprv1alpha1.TalosUpgradeList{}
-		if err := c.List(ctx, talosUpgrades); err != nil {
-			return false, "", fmt.Errorf("failed to list TalosUpgrade resources: %w", err)
-		}
-
-		for _, upgrade := range talosUpgrades.Items {
-			if upgrade.Name == currentUpgradeName {
-				continue
-			}
-
-			// if two plans started at the exact same second.
-			// The controller's natural Reconcile order resolves the tie-break.
-			if upgrade.Status.Phase == constants.PhaseInProgress {
-				return true, fmt.Sprintf("Waiting for another TalosUpgrade plan '%s' to complete", upgrade.Name), nil
-			}
-		}
-	}
-
-	return false, "", nil
-}
