+++ internal/webhook/kubernetesupgrade/webhook.go
@@ -1,4 +1,4 @@
-package webhook
+package kubernetesupgrade
 
 import (
 	"context"
@@ -11,11 +11,12 @@ import (
 
 	tupprv1alpha1 "github.com/home-operations/tuppr/api/v1alpha1"
 	"github.com/home-operations/tuppr/internal/constants"
+	"github.com/home-operations/tuppr/internal/webhook/validation"
 )
 
 var kuberneteslog = logf.Log.WithName("kubernetes-resource")
 
-type KubernetesUpgradeValidator struct {
+type Validator struct {
 	Client            client.Client
 	TalosConfigSecret string
 	Namespace         string
@@ -23,27 +24,26 @@ type KubernetesUpgradeValidator struct {
 
 // +kubebuilder:webhook:path=/validate-tuppr-home-operations-com-v1alpha1-kubernetesupgrade,mutating=false,failurePolicy=fail,sideEffects=None,groups=tuppr.home-operations.com,resources=kubernetesupgrades,verbs=create;update,versions=v1alpha1,name=vkubernetesupgrade.kb.io,admissionReviewVersions=v1
 
-var _ admission.Validator[*tupprv1alpha1.KubernetesUpgrade] = &KubernetesUpgradeValidator{}
+var _ admission.Validator[*tupprv1alpha1.KubernetesUpgrade] = &Validator{}
 
 // ValidateCreate implements webhook.CustomValidator so a webhook will be registered for the type
-func (v *KubernetesUpgradeValidator) ValidateCreate(ctx context.Context, k *tupprv1alpha1.KubernetesUpgrade) (admission.Warnings, error) {
+func (v *Validator) ValidateCreate(ctx context.Context, k *tupprv1alpha1.KubernetesUpgrade) (admission.Warnings, error) {
 	kuberneteslog.Info("validate create", "name", k.Name, "namespace", k.Namespace, "version", k.Spec.Kubernetes.Version)
 	return v.validate(ctx, k)
 }
 
 // ValidateUpdate implements webhook.CustomValidator so a webhook will be registered for the type
-
-func (v *KubernetesUpgradeValidator) ValidateUpdate(ctx context.Context, old, k *tupprv1alpha1.KubernetesUpgrade) (admission.Warnings, error) {
+func (v *Validator) ValidateUpdate(ctx context.Context, old, k *tupprv1alpha1.KubernetesUpgrade) (admission.Warnings, error) {
 	kuberneteslog.Info("validate update", "name", k.Name)
 
-	if err := ValidateUpdateInProgress(old.Status.Phase, old.Spec, k.Spec); err != nil {
+	if err := validation.ValidateUpdateInProgress(old.Status.Phase, old.Spec, k.Spec); err != nil {
 		return nil, err
 	}
 	return v.validate(ctx, k)
 }
 
 // ValidateDelete implements webhook.CustomValidator so a webhook will be registered for the type
-func (v *KubernetesUpgradeValidator) ValidateDelete(ctx context.Context, k *tupprv1alpha1.KubernetesUpgrade) (admission.Warnings, error) {
+func (v *Validator) ValidateDelete(ctx context.Context, k *tupprv1alpha1.KubernetesUpgrade) (admission.Warnings, error) {
 	if k.Status.Phase == constants.PhaseInProgress {
 		return admission.Warnings{
 			fmt.Sprintf("Deleting KubernetesUpgrade '%s' while upgrade is in progress. This may leave the cluster in an inconsistent state.", k.Name),
@@ -52,39 +52,39 @@ func (v *KubernetesUpgradeValidator) ValidateDelete(ctx context.Context, k *tupp
 	return nil, nil
 }
 
-func (v *KubernetesUpgradeValidator) validate(ctx context.Context, k *tupprv1alpha1.KubernetesUpgrade) (admission.Warnings, error) {
+func (v *Validator) validate(ctx context.Context, k *tupprv1alpha1.KubernetesUpgrade) (admission.Warnings, error) {
 	var warnings admission.Warnings
 
 	list := &tupprv1alpha1.KubernetesUpgradeList{}
-	if err := ValidateSingleton(ctx, v.Client, "KubernetesUpgrade", k.Name, list); err != nil {
+	if err := validation.ValidateSingleton(ctx, v.Client, "KubernetesUpgrade", k.Name, list); err != nil {
 		return warnings, err
 	}
 
-	if _, err := ValidateTalosConfigSecret(ctx, v.Client, v.TalosConfigSecret, v.Namespace); err != nil {
+	if _, err := validation.ValidateTalosConfigSecret(ctx, v.Client, v.TalosConfigSecret, v.Namespace); err != nil {
 		return warnings, err
 	}
 
-	if err := ValidateVersionFormat(k.Spec.Kubernetes.Version); err != nil {
+	if err := validation.ValidateVersionFormat(k.Spec.Kubernetes.Version); err != nil {
 		return warnings, fmt.Errorf("invalid kubernetes version: %w", err)
 	}
 
-	if err := ValidateHealthChecks(k.Spec.HealthChecks); err != nil {
+	if err := validation.ValidateHealthChecks(k.Spec.HealthChecks); err != nil {
 		return warnings, err
 	}
 
-	if err := ValidateTalosctlSpec(k.Spec.Talosctl); err != nil {
+	if err := validation.ValidateTalosctlSpec(k.Spec.Talosctl); err != nil {
 		return warnings, err
 	}
 
 	// Validate maintenance window if specified
-	if mwWarnings, err := validateMaintenanceWindows(k.Spec.Maintenance); err != nil {
+	if mwWarnings, err := validation.ValidateMaintenanceWindows(k.Spec.Maintenance); err != nil {
 		return warnings, fmt.Errorf("spec.maintenanceWindow validation failed: %w", err)
 	} else {
 		warnings = append(warnings, mwWarnings...)
 	}
 
 	// Add warnings for risky configurations
-	warnings = append(warnings, GenerateCommonWarnings(
+	warnings = append(warnings, validation.GenerateCommonWarnings(
 		k.Spec.Kubernetes.Version,
 		k.Spec.HealthChecks,
 		k.Spec.Talosctl.Image.Tag,
@@ -94,7 +94,7 @@ func (v *KubernetesUpgradeValidator) validate(ctx context.Context, k *tupprv1alp
 	return warnings, nil
 }
 
-func (v *KubernetesUpgradeValidator) SetupWebhookWithManager(mgr ctrl.Manager) error {
+func (v *Validator) SetupWebhookWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewWebhookManagedBy(mgr, &tupprv1alpha1.KubernetesUpgrade{}).
 		WithValidator(v).
 		Complete()
